requirements/base.txt
Django==2.0.7

python-dotenv==0.8.2
hxlti-djapp==0.1.1
